% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lookup.R
\name{add_lookup}
\alias{add_lookup}
\title{Add a geographic facet lookup to a database}
\usage{
add_lookup(db, df)
}
\arguments{
\item{db}{A SWIM scenario database}

\item{df}{A lookup table with two colums:
\enumerate{
  \item{A beta zone field \code{BZONE}, depending on the level of the zone
  aggregation.}
  \item{An aggregation field, which will have some value for all zones. The
  name of the data column will be the name of field in the table.}
}}
}
\value{
Writes the new column to the \code{ALLZONES} table of the database.
}
\description{
Add a geographic facet lookup to a database
}
