% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athlete_query.R
\name{athlete_count}
\alias{athlete_count}
\title{Query For Counting Athletes Per Team}
\usage{
athlete_count(con, team_name)
}
\arguments{
\item{con}{A database connection.}

\item{team_name}{A character vector with the name of an NCAA swimming conference.}
}
\value{
a data frame with the count on athletes by gender
}
\description{
Get a count of athletes by team.
}

