% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calc_exprate}
\alias{calc_exprate}
\title{Calculate exponential growth rate}
\usage{
calc_exprate(p1, p2, t1, t2)
}
\arguments{
\item{p1}{Value in period 1}

\item{p2}{Value in period 2}

\item{t1}{Time step for period 1 (i.e., 2010)}

\item{t2}{Time step for period 2 (i.e., 2040)}
}
\description{
Calculate exponential growth rate
}
\details{
This function calculates an annualized exponential growth rate
  implied by measuring a value in two different time periods. Solves the
  elementary exponential growth equation, \deqn{p_2 = p_1 e^{r(t_2 - t_1)}}
  for \eqn{r}.
}
\examples{
calc_exprate(10, 12, 2010, 2040)
}
