% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calc_grt}
\alias{calc_grt}
\title{Calculate the growth rate in a vector}
\usage{
calc_grt(x)
}
\arguments{
\item{x}{A numeric vector in the proper order}
}
\value{
A numeric vector of length(x)
}
\description{
Calculate the growth rate in a vector
}
