% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calc_index}
\alias{calc_index}
\title{Calculate the indexed value of a vector}
\usage{
calc_index(x)
}
\arguments{
\item{x}{A numeric vector, in the proper order}
}
\value{
A numeric vector of length(x)
}
\description{
Calculate the indexed value of a vector
}

