% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone_leaflet.R
\name{change_ggmap}
\alias{change_ggmap}
\title{Compare socioeconomic variables on a ggmap}
\usage{
change_ggmap(db1, db2, scope = NULL, year = 2010, metric = c("diff",
  "pct"), variable = c("Population", "Employment", "HH"),
  scen_names = c("Reference", "Current"), ggmap = FALSE)
}
\description{
Compare socioeconomic variables on a ggmap
}

