% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone_leaflet.R
\name{change_leaflet}
\alias{change_leaflet}
\title{Leaflet plot of change over time in one scenario.}
\usage{
change_leaflet(db, year1 = 2010, year2 = 2030)
}
\arguments{
\item{db}{The scenario database}

\item{year1}{T}
}
\description{
Leaflet plot of change over time in one scenario.
}
\details{
This function creates an interactive Leaflet plot of the alpha zone
  system, showing annualized implied growth rates of key socioeconomic
  variables between two years. For instance, the user can select 2010 and
  2030 as analysis years, and the plot will show the implied average
  exponential growth rate between them.

  As a note, SWIM does not generate alpha-zone level statistics in years when
  the transport model does not run. If the user asks for year that does not
  exist, then the plot will use the nearest available year.
}

