% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dates.R
\name{clean_swim_date}
\alias{clean_swim_date}
\title{Clean up dates in USA Swimming data download}
\usage{
clean_swim_date(x)
}
\arguments{
\item{x}{a data frame in the format downloaded from USA swimmings top times report.}
}
\description{
Clean up dates
}
\keyword{internal}
