% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employment.R
\name{compare_employment}
\alias{compare_employment}
\title{Compare labor output between two scenarios}
\usage{
compare_employment(db1, db2, ...)
}
\arguments{
\item{db1}{The reference scenario database.}

\item{db2}{The current scenario database.}
}
\value{
A ggplot2 object showing the employment by type and and year.
}
\description{
This function compares the employment by type in an area over time between two
scenarios.
}

