% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdp.R
\name{compare_gdp}
\alias{compare_gdp}
\title{Compare GDP between two scenarios}
\usage{
compare_gdp(db1, db2, facet_var = c("MPO", "COUNTY", "STATE"),
  facet_levels = NULL, color_levels = NULL)
}
\arguments{
\item{db1}{The swim database for the "Reference" scenario.}

\item{db2}{The swim database for the "Current" scenario.}

\item{facet_var}{Field to facet by: "MPO", "COUNTY", or "STATE".}

\item{facet_levels}{A character vector of the variable specifiying
which levels to include.}
}
\value{
ggplot2 object
}
\description{
Compare GDP between two scenarios
}

