% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logsums.R
\name{compare_logsums}
\alias{compare_logsums}
\title{Compare logsums from two scenarios}
\usage{
compare_logsums(
  db1,
  db2,
  color_var = c("COUNTY", "MPO", "ALDREGION", "STATE", "DOT_REGION"),
  color_levels = NULL,
  index_year = 2000
)
}
\arguments{
\item{db1}{The reference scenario VIZ database.}

\item{db2}{The current scenario VIZ database.}

\item{color_var}{The variable to color the plot by.}

\item{color_levels}{The variable to}

\item{index_year}{[Optional] index year that should be used as the starting year for data or plots;}
}
\description{
Compare logsums from two scenarios
}
