% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floorspace.R
\name{compare_occupancy}
\alias{compare_occupancy}
\title{Compare occupancy in a scenario over time.}
\usage{
compare_occupancy(db1, db2, ...)
}
\arguments{
\item{db1}{The swim database for the "Reference" scenario.}

\item{db2}{The swim database for the "Current" scenario.}

\item{...}{Arguments passed on to \code{extract_rents}
\describe{
  \item{db}{The scenario database.}
  \item{facet_var}{The variable in the zone table to facet by. Defaults to MPO}
  \item{facet_levels}{The levels of the facet variable to keep. Defaults to all
levels other than external stations.}
  \item{type_levels}{The types of employment to show in the plot.}
  \item{index}{Should the function extract indexed or absolute values?}
}}
}
\value{
A ggplot2 plot object
}
\description{
Compare occupancy in a scenario over time.
}

