% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vmt.R
\name{compare_pct_cong}
\alias{compare_pct_cong}
\title{Plot percent congested links over time.}
\usage{
compare_pct_cong(db1, db2, facet_var = "MPO", facet_levels = NULL,
  congested_voc = 0.9)
}
\arguments{
\item{db1}{The swim database for the "Reference" scenario.}

\item{db2}{The swim database for the "Current" scenario.}

\item{facet_var}{The variable to use in faceting the data.}

\item{facet_levels}{A character vector giving the levels of the faceting
  variable to use. Levels not called are dropped from the plot; default is
  \code{NULL}, meaning print all level.}
}
\value{
A ggplot2 figure object.
}
\description{
This function plots the percent of congested links, colored and
faceted by arbitrary variables in the link table.
}

