% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/population.R
\name{compare_sevar}
\alias{compare_sevar}
\title{Compare population and employment}
\usage{
compare_sevar(db1, db2, facet_var = c("MPO", "COUNTY"), facet_levels = NULL)
}
\arguments{
\item{db1}{The swim database for the "Reference" scenario.}

\item{db2}{The swim database for the "Current" scenario.}

\item{facet_var}{Field to facet by: either "MPO" or "COUNTY".}

\item{facet_levels}{A character vector of the facet variable specifiying
  which levels to include.}
}
\value{
A \code{ggplot2} plot object showing the modeled change in employment
  and population over time.
}
\description{
Compare population and employment
}

