% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tlfd.R
\name{compare_tlfd}
\alias{compare_tlfd}
\title{Compare Trip Length Frequency Distributions}
\usage{
compare_tlfd(db1, db2, region_var = c("MPO", "COUNTY", "STATE"),
  regions = NULL, bin_width = 1, max_bin = 50)
}
\arguments{
\item{db1}{The swim database for the "Reference" scenario.}

\item{db2}{The swim database for the "Current" scenario.}

\item{region_var}{The region to aggregate to.}

\item{regions}{The regions to return.  If \code{NULL}, returns all.}

\item{bin_width}{The width of bins to use in the trip length frequency
  distribution. Defaults to 1.}
}
\value{
A ggplot2 object.
}
\description{
Compare Trip Length Frequency Distributions
}

