% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmt.R
\name{compare_vmt}
\alias{compare_vmt}
\title{Compare VMT between scenarios}
\usage{
compare_vmt(db1, db2, facet_var = c("MPO", "COUNTY"), facet_levels = NULL)
}
\arguments{
\item{db1}{The swim database for the "Reference" scenario.}

\item{db2}{The swim database for the "Current" scenario.}

\item{facet_var}{Field to facet by: either "MPO" or "COUNTY".}

\item{facet_levels}{A character vector of the facet variable specifiying
which levels to include.}
}
\value{
A ggplot2 object showing VMT by facility type in each facet level
  over time.
}
\description{
Compare VMT between scenarios
}

