% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laborforce.R
\name{compare_wapr}
\alias{compare_wapr}
\title{Compare workforce participation in two scenarios}
\usage{
compare_wapr(db1, db2, facet_var = c("BZONE", "MPO", "COUNTY",
  "DOT_REGION", "STATE"), facet_levels = NULL)
}
\arguments{
\item{db1}{The swim database for the "Reference" scenario.}

\item{db2}{The swim database for the "Current" scenario.}

\item{facet_var}{Field to facet by: either "MPO" or "COUNTY".}

\item{facet_levels}{A character vector of the facet variable specifiying
which levels to include.}
}
\value{
ggplot2 object
}
\description{
Compare workforce participation in two scenarios
}
