% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTables.R
\name{create_event}
\alias{create_event}
\title{Code to create each table in the database.}
\usage{
create_event(con)
}
\arguments{
\item{con}{is a database connection object of class \link[DBI]{
DBIConnection-class}}
}
\description{
Code to create each table in the database.
}
\keyword{internal}
