\name{createFrequencyMatrix}
\alias{createFrequencyMatrix}

\title{
 Analyzing the frequency of worm thrashing over time and creating 
 the frequency matrix and annotation file
}

\description{
 This function analyzes the frequency of worm thrashing over time 
 for the files returned by the Tracker program and then combined 
 the frequency of all files into a matrix and extracted genotype 
 information from tracker file names to create annotation file.
}

\arguments{
  \item{inputPath}{
    \code{inputPath} is a directory which contains the files returned by 
 the Tracker program. Because annotation file is generated by extracting 
 the genotype information from tracker file names, the user should use dashes 
 to denote genotypes and separate the date in the file name like 
 Genotype_Drug(if used)_Dose(if used)_Date_#. The following are some examples 
 of tracker file names: N4_AMPH_100uM_2-11-14_1 (genotype is N4), 
 dat-1ok157_IMI_10uM_12-12-12_7(genotype is dat-1ok157), 
 cat-2e1112dat-1ok157_2-20-09_2 (genotype is cat-2e1112dat-1ok157) 
 and dat-1ok157dop-3vs106_5-17-06_4 (genotype is dat-1ok157dop-3vs106).
  }
  \item{outputPath}{
    \code{outputPath} is a directory which saves the plots and 
 files returned by the function.
  }
  \item{method}{
    The function provides four different counting methods: 
 "FFT" (Fast Fourier Transform), "Extrema", "PeakDet" (peak delta) 
 and "RT+GP" (Get Peaks plus Racetrack Filter) and the users can 
 select one of them to output the corresponding frequency matrix. 
 The default outputted \code{method} is "Extrema".
  }
  \item{Threshold}{
    \code{Threshold} is the amount of degrees (in radians) required 
 to count at as thrash and the default is 0.6.
  }
  \item{DeltaPeakDt}{
    \code{DeltaPeakDt} is the threshold for "peak delta" algorithm, 
 similar to \code{Threshold} and the default is 1.6.
  }
  \item{MinFrameBtwnMax}{
    \code{MinFrameBtwnMax} is the minimum number of frames 
 between maxima and the default is 4.
  }
  \item{MinDelta}{
    \code{MinDelta} is similar to \code{DeltaPeakDet} 
 and the default is 2.5.
  }
  \item{longPeriod}{
    \code{longPeriod} is the longest period cycle which is not zerod 
 and the default is 5.
  }
  \item{AvWindowSize}{
    \code{AvWindowSize} is the length of the average window in seconds 
 and the default is 10 seconds.
  }
  \item{fps}{
    \code{fps} is the frame per second and the default is 15.
  }
  \item{ZP_Length}{
    \code{ZP_Length} is the Zero-padding length and the default is 100.
  }
  \item{WindowSize}{
    \code{WindowSize} is the size of window for computing the Fast Fourier Transform 
 and the default is 30.
  }
  \item{MaxCompWin}{
    \code{MaxCompWin} is the window size on deciding if the current angle is a maxima 
 and the default is 2.
  }
  \item{minTime}{
    \code{minTime} is the minimum threshold of time points for the following analysis
 and the default is 0 second.
  }
  \item{maxTime}{
    \code{maxTime} is the maximum threshold of time points for the following analysis 
 and the default is 600 seconds.
  }
}

\value{
  The createFrequencyMatrix function returns a list object which contains the following information:
  \item{frequencyMatrix}{
    \code{frequencyMatrix} combines the analysis results 
 of the frequency of worm thrashing over time for all Tracker files in the \code{inputPath}.
  }
  \item{annotation}{
    \code{annotation} contains all genotype information extracted 
 from file names of all Tracker files in the \code{inputPath}.
  }
}

\details{
  The createFrequencyMatrix function outputs six files:
   
  1. \code{outputDescription_createFrequencyMatrix.html} contains a summary of 
 all output files.

  2. \code{XFig.jpg} is the image of scatter plot of one animal plotted as 
 "Frequency vs Time(min)" with all four counting methods overlaid. 
 "X" of "XFig.jpg" represents the input file names.

  3. \code{XFigSub.jpg} is the same as \code{XFig.jpg} except counting methods 
 are broken up into four different plots. This is very helpful in checking through 
 a video to make sure that Tracker tracked the worm properly. Bad contrast can be 
 a problem with Tracker missing the worm and these files help to identify troublesome 
 videos for retracking or discarding.

  4. \code{XFreq.csv} is the CSV file of raw data organized by column, 
 where column one represents frequency as counted by FFT, column two represents 
 frequency calculated by Extrema, column three represents frequency calculated by 
 PeakDt, column four represents frequency as counted by RT+GP and column five 
 represents time in seconds.

  5. \code{frequencyMatrix.txt} is a TXT file which contains the information of \code{frequencyMatrix}.

  6. \code{annotationfile.txt} is a TXT file which contains the information of \code{annotation}.
}

\author{
    Jing Wang and Andrew Hardaway
}

\note{
    The worm frequency analysis script in this function is converted from matlab script 
 coded by Katherine Fleming and Paul Fleming.
}

\examples{
    inputPath <- system.file("extdata","trackerFiles",package="SwimR")
    outputPath <- getwd()
    freMat <- createFrequencyMatrix(inputPath, outputPath, method = "Extrema", 
 Threshold = 0.6, DeltaPeakDt = 1.6, MinFrameBtwnMax = 4, MinDelta = 2.5, longPeriod = 5, 
 AvWindowSize = 10, fps = 15, ZP_Length = 100, WindowSize = 30, MaxCompWin = 2, minTime = 0, 
 maxTime = 600)
}

\seealso{
    \code{\link{SwimR}}
}

\keyword{ methods }

