% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{cut_abserror}
\alias{cut_abserror}
\title{Cut absolute differences into ranges}
\usage{
cut_abserror(x)
}
\arguments{
\item{x}{a vector of error measurements}
}
\value{
a factor showing the bin
}
\description{
Cut absolute differences into ranges
}
