% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{cut_diverror}
\alias{cut_diverror}
\title{Cut divergins percent error into ranges}
\usage{
cut_diverror(x)
}
\arguments{
\item{x}{a vector of error measurements}
}
\value{
a factor showing the bin
}
\description{
Cut divergins percent error into ranges
}

