% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{cut_grt}
\alias{cut_grt}
\title{Cut growth rates into ranges}
\usage{
cut_grt(x)
}
\arguments{
\item{x}{a vector of error measurements}
}
\value{
a factor showing the bin
}
\description{
Cut growth rates into ranges
}
