% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{D3Qualifying}
\alias{D3Qualifying}
\title{NCAA Qualifying Times}
\format{A data frame with 52 rows and 5 columns:
\describe{
\item{event}{}
\item{gender}{}
\item{standard}{}
 \item{swim_time}{}
 \item{swim_time2}{}
}}
\source{
\url{https://swimswam.com/ncaa-releases-division-iii-time-standards/}
}
\usage{
D3Qualifying
}
\description{
A Dataset containing the qualifying times for the 2017 NCAA Division II 
Swimming Championships.
}
\keyword{datasets}

