% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_load.R
\name{db_load}
\alias{db_load}
\title{Load Swimming Results Into A Database}
\usage{
db_load(db_name, data, team)
}
\arguments{
\item{db_name}{A character string providing the filename of the SQLite Database.}

\item{data}{A data frame in the format of a USA Swimming Top Times Report.}

\item{team}{A character string providing the team name for the team of interest.  The team name coresponds to the table name.}
}
\description{
Create and/or Load Data into a SQLite Database
}
\details{
\code{db_load} is used to create and load a USA Swimmings Top Times Report into  SQLite database.  The function utilizes \code{\link[DBI]{DBI}} create the connection to the database, creation of the tables and insertion of the data.

The function is designed so that each team has its own table.  The team name is provided in the function call.  If the table does not exist it is created and the data is loaded.  If the table does exist the data is appended to the existing table.  The function currently does not check for duplicate entries.

In version of \strong{\code{swimR}}, the database is not designed around a particular schema.  primary keys are not currently implemented.  The tables are not currently set to be joined, but rather unioned.
}

