% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone_leaflet.R
\name{diff_popup}
\alias{diff_popup}
\title{Build diff popup tag for leaflet}
\usage{
diff_popup(shp, var, scen_names)
}
\arguments{
\item{shp}{An object of class \code{SpatialPolygonsDataFrame} with the
plotting fields already appended.}

\item{var}{Variable to create popup tag for (e.g., \code{"pop"})}

\item{scen_names}{Names of the scenarios in the comparison. Defaults to
"Reference", "Current".}
}
\description{
Build diff popup tag for leaflet
}

