% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{discover_outlying_rates}
\alias{discover_outlying_rates}
\title{Discover outlying counties}
\usage{
discover_outlying_rates(db, counties = NULL, method = c("count", "pval"))
}
\arguments{
\item{db}{The swim database}

\item{counties}{A character vector of counties to calculate.}

\item{method}{The method by which unreasonable counties are identified. See
"Details".}
}
\value{
A data frame with information on each county's conformity with
  historical trends.
}
\description{
A major question we face is identifying whether the modeled population growth
rate in a region is "reasonable." To date reasonableness has been subjective,
but this function provides quantitative methodologies to calculate this.
}
\details{
If \code{method = "count"} (default), the number of years for which
  the annualized population growth rate lies outside of the historical
  confidence interval. If \code{method = "pval"}, a data frame of swim scenario
  years with with the $p$-value of the annualized growth rate on the
  distribution defined by historical data.
}
\seealso{
\code{\link{pop_rate}}
}
