% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth_queries.R
\name{event_depth}
\alias{event_depth}
\title{Count Number Of Athletes Competing In Each Event.}
\usage{
event_depth(con, team_name, gender = "Both")
}
\arguments{
\item{con}{a database connection.}

\item{team_name}{A character vector with the name of an NCAA swimming
team.}

\item{gender}{a character vector.  Indicates the output should include men,
women or both.  Defaults to "Both".  Other options are "M" for men and "F"
for women.}
}
\value{
A data frame.
}
\description{
The function queries the database to determine the number of athletes swimming
a given event.  Aggregation is done by event and gender.
}
\examples{
df <- event_depth(con, team_name = 'Iowa', gender = 'Both')

}
