% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_naming.R
\name{event_naming}
\alias{event_naming}
\title{Create Clean Event Names}
\usage{
event_naming(data)
}
\arguments{
\item{data}{A data frame passed from \code{data_clean}.}
}
\description{
An Internal Function used in \code{\link{data_clean}}
}
\keyword{internal}
