% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone_leaflet.R
\name{extract_changedata}
\alias{extract_changedata}
\title{Extract data for leaflet zone plots}
\usage{
extract_changedata(db, year1, year2)
}
\arguments{
\item{db}{The scenario database}

\item{year1}{T}
}
\description{
Extract data for leaflet zone plots
}

