% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floorspace.R
\name{extract_floorspace}
\alias{extract_floorspace}
\title{Extract floorspace statistics from scenario}
\usage{
extract_floorspace(
  db,
  facet_var = c("MPO", "COUNTY", "STATE"),
  facet_levels = NULL,
  type_levels = NULL,
  index = TRUE,
  index_year = 2000
)
}
\arguments{
\item{db}{The scenario database.}

\item{facet_var}{The variable in the zone table to facet by. Defaults to MPO}

\item{facet_levels}{The levels of the facet variable to keep. Defaults to all
levels other than external stations.}

\item{type_levels}{The types of employment to show in the plot.}

\item{index}{Should the function extract indexed or absolute values?}

\item{index_year}{[Optional] index year that should be used as the starting year for data or plots;}
}
\description{
This is an internal function to pull the floorspace data from a scenario and
return it to either the single scenario or scenario comparison functions.
}
