% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdp.R
\name{extract_gdp}
\alias{extract_gdp}
\title{Extract total wages}
\usage{
extract_gdp(db, facet_var = c("MPO", "COUNTY", "STATE"),
  facet_levels = NULL, color_levels = NULL)
}
\arguments{
\item{db}{The scenario sqlite database.}

\item{facet_var}{Field to facet by: "MPO", "COUNTY", or "STATE".}

\item{facet_levels}{A character vector of the variable specifiying
which levels to include.}
}
\value{
A \code{data_frame} with the participation rate in each facet region
  in each transport model year.
}
\description{
This function returns the total dollar value of all the labor sold by a
collection of zones over time.
}

