% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/laborforce.R
\name{extract_lfpr}
\alias{extract_lfpr}
\title{Extract labor force participation}
\usage{
extract_lfpr(db, facet_var = c("BZONE", "MPO", "COUNTY", "DOT_REGION",
  "STATE"), facet_levels = NULL, color_levels = NULL)
}
\arguments{
\item{db}{The scenario sqlite database.}

\item{facet_var}{Field to facet by: "MPO", "COUNTY", or "STATE".}

\item{facet_levels}{A character vector of the variable specifiying
  which levels to include.}
}
\value{
A \code{data_frame} with the participation rate in each facet region
  in each transport model year.
}
\description{
BLS defines the civilian labor force participation rate as the percent of
individuals age 16 and over who are either working or who are looking for
work. This is not something we ask about in travel surveys generally, so
it is not possible to make an apples-to-apples comparison.
}
\details{
This function returns the number of workers in each facet region divided
by the number of individuals fifteen years and older in that region. The
definition of a "worker" in a travel model is different than the BLS definition
in that travel modesl do not consider unemployed individuals who are looking
for work.
}

