% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logsums.R
\name{extract_logsums}
\alias{extract_logsums}
\title{Extract destination choice logsums}
\usage{
extract_logsums(
  db,
  scope = NULL,
  purposes = NULL,
  agg_var = "AZONE",
  index_year = 2000
)
}
\arguments{
\item{db}{The scenario sqlite database.}

\item{scope}{a dplyr::filtering criteria to limit the scope of the dataframe}

\item{purposes}{a vector of trip purposes to include in the average logsum.}

\item{agg_var}{The region variable on which to aggregate logsums.}

\item{index_year}{[Optional] index year that should be used as the starting year for data or plots;}
}
\value{
a data_frame
}
\description{
The logsums are provided at production-side beta zones, by purpose and market
segment. This function aggregates market segments by averaging and purposes
by summation (sum of all purpose logsums faced by the average user).
}
