% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelwide.R
\name{extract_modelwide}
\alias{extract_modelwide}
\title{Extract model-wide composite utilities}
\usage{
extract_modelwide(db, filters)
}
\arguments{
\item{db}{Scenario database.}

\item{filters}{list object containing character strings on which to dplyr::filter
parameters.}
}
\description{
Extract model-wide composite utilities
}

