% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floorspace.R
\name{extract_rents}
\alias{extract_rents}
\title{Extract rent price}
\usage{
extract_rents(db, facet_var = c("MPO", "COUNTY", "STATE"),
  facet_levels = NULL, type_levels = NULL, index = TRUE)
}
\arguments{
\item{db}{The scenario database.}

\item{facet_var}{The variable in the zone table to facet by. Defaults to MPO}

\item{facet_levels}{The levels of the facet variable to keep. Defaults to all
levels other than external stations.}

\item{type_levels}{The types of employment to show in the plot.}

\item{index}{Should the function extract indexed or absolute values?}
}
\description{
This function extracts the occupancy rate by floortype from a scenario over
time.
}

