% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{extract_se}
\alias{extract_se}
\title{Extract SE data from DB}
\usage{
extract_se(db, color_var = c("MPO", "COUNTY"), color_levels = NULL,
  controls = TRUE, index = FALSE)
}
\arguments{
\item{db}{The scenario sqlite database.}

\item{color_var}{Field to color by: either "MPO" or "COUNTY".}

\item{color_levels}{A character vector of the color variable specifiying
which levels to include.}

\item{controls}{Plot against the control totals. Defaults to TRUE, cannot
currently run with FALSE.}

\item{index}{Whether to show the variables as indexed against the base year.}
}
\value{
a data frame
}
\description{
Extract SE data from DB
}

