% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlfd.R
\name{extract_tlfd}
\alias{extract_tlfd}
\title{Extract trip length frequency distribution data}
\usage{
extract_tlfd(
  db,
  region_var = c("MPO", "COUNTY", "STATE"),
  regions = NULL,
  cumulative = FALSE,
  years = NULL,
  index_year = 2000
)
}
\arguments{
\item{db}{The scenario database.}

\item{region_var}{The region to aggregate to.}

\item{regions}{The regions to return.  If \code{NULL}, returns all.}

\item{cumulative}{Return a cumulative PDF?}

\item{years}{The years to show in the plot. If null, will show the first}

\item{index_year}{[Optional] index year that should be used as the starting year for data or plots;}

\item{max_bin}{The maximum bin size to return.}
}
\value{
A data frame with the trip length frequency distribution by year
  and region.
}
\description{
Extract trip length frequency distribution data
}
