% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vmt.R
\name{extract_vht}
\alias{extract_vht}
\title{Extract VHT}
\usage{
extract_vht(db, facet_var, facet_levels)
}
\arguments{
\item{db}{The scenario database.}

\item{facet_var}{The variable to use in faceting the data.}

\item{facet_levels}{A character vector giving the levels of the faceting
  variable to use. Levels not called are dropped from the plot; default is
  \code{NULL}, meaning print all level.}
}
\value{
A data frame with the
}
\description{
Extract VHT
}

