% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmt.R
\name{extract_vmt}
\alias{extract_vmt}
\title{Extract VMT}
\usage{
extract_vmt(db, facet_var = "MPO", facet_levels = NULL, index = FALSE)
}
\arguments{
\item{db}{The scenario database.}

\item{facet_var}{The variable to use in facetting the data.}

\item{facet_levels}{A character vector giving the levels of the facet
variable to use. Levels not called are dropped from the plot; default is
\code{NULL}, meaning print all levels.}

\item{index}{Whether to show the variables as indexed against the base year.}
}
\value{
A data frame with VMT summarized by facility type and facet variable.
}
\description{
Extract VMT
}

