% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone_leaflet.R
\name{extract_zonedata}
\alias{extract_zonedata}
\title{Extract se data for leaflet zone plots}
\usage{
extract_zonedata(db, year1, year2 = NULL)
}
\arguments{
\item{db}{The scenario database}

\item{year1}{The first year}

\item{year2}{The second year}
}
\description{
Extract se data for leaflet zone plots
}
\details{
If \code{year2 = NULL}, then will return only the values from year 1
  and will not calculate the implied exponential growth rate.
}
