% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fac_types}
\alias{fac_types}
\title{Facility types}
\format{A data frame with 17 observations.
\describe{
  \item{PLANNO}{The facility classifications in the db.}
  \item{FacType}{The consolidated classifications, as an ordered factor.}
}}
\usage{
fac_types
}
\description{
A dataset mapping the facility types in the database to consolidated
categories.
}
\keyword{datasets}

