% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{floor_types}
\alias{floor_types}
\title{Floor types}
\format{A data frame with 16 observations.
\describe{
  \item{commodity}{The floorspace commodity types in the db.}
  \item{floor_type}{The consolidated categories.}
}}
\usage{
floor_types
}
\description{
A dataset mapping the floor types in the database to consolidated categories.
}
\keyword{datasets}
