% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commute_times.R
\name{get_commute_times}
\alias{get_commute_times}
\title{Calculate trip times by residence region}
\usage{
get_commute_times(dir, year, region_var = c("MPO", "COUNTY", "STATE"),
  regions = NULL)
}
\arguments{
\item{dir}{Scenario directory}

\item{year}{Analysis year, e.g. \code{t20}.}

\item{region_var}{One of \code{c("MPO", "COUNTY", "STATE")}}

\item{regions}{A list of regions to include; defaults to \code{NULL} (all).}
}
\description{
Calculate trip times by residence region
}
\details{
Note that this function does not work on a VIZ database, but rather
  on the scenario outputs directly.
}
