% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_press.R
\name{get_conferences}
\alias{get_conferences}
\title{Get a list of college swimming confernces.}
\usage{
get_conferences(division, printer = FALSE)
}
\arguments{
\item{division}{integer.  1 for NCAA Division I. 2 for NCAA Division II, 3 for NCAA Divison III.}

\item{printer}{logical.  If TRUE the list of conferences will be printed to the console.  Defaults to FALSE.}
}
\value{
a data frame.
}
\description{
Get a list of college swimming confernces.
}
\details{
The function collects college swimming conferences.  The list is generated by parsing the html behind the confernce drop down menu on \href{http://www.usaswimming.org/DesktopDefault.aspx?TabId=1971&Alias=Rainbow&Lang=en}{USA Swimming's Top Times Report}.

The list of college swimming confernces is returned as a one column data frame.
}

