% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_validation.R
\name{get_counts}
\alias{get_counts}
\title{Get counts table}
\usage{
get_counts(db, trucks = FALSE)
}
\arguments{
\item{db}{The scenario database.}

\item{trucks}{Get truck counts or AAWDT counts?}
}
\value{
A counts dataframe w
}
\description{
Get counts table
}

