% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_levels}
\alias{get_levels}
\title{Return all levels of a factor other than external stations}
\usage{
get_levels(df)
}
\arguments{
\item{df}{a tbl_sqlite with a faceting variable named \code{facet_var}}
}
\value{
names of all levels of the factor
}
\description{
Return all levels of a factor other than external stations
}
