% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim_display.R
\name{get_skim}
\alias{get_skim}
\title{Process skim values from one zone.}
\usage{
get_skim(skim, zone, from = TRUE)
}
\arguments{
\item{skim}{Path to zmx skim file}

\item{zone}{Focus zone for the skim calculation.}

\item{from}{If \code{TRUE}, gets values \em{from} \code{zone} to other zones.}
}
\description{
Process skim values from one zone.
}

