% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{historical_pop}
\alias{historical_pop}
\title{Historical county population}
\format{
A data frame with 324 observations.
\describe{
  \item{county}{name of the county in Oregon}
  \item{year}{1930-2010 by decennial.}
  \item{y}{The population count (coded as y for joining to interior data).}
}
}
\source{
Reconstructed from data provided by Rick Donnelly
}
\usage{
historical_pop
}
\description{
A dataset containing the historical population from the decennial census of
population and housing for Oregon from 1930-2010
}
\keyword{datasets}
