% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_create.R
\name{id_create}
\alias{id_create}
\title{Create Unique ID for each athlete.}
\usage{
id_create(data)
}
\arguments{
\item{data}{The data frame provided in \code{\link{data_clean}}}
}
\value{
The function creates a new column
}
\description{
An internal function used by \code{\link{data_clean}}
}
\details{
\code{id_create} is an internal function used by \code{\link{data_clean}}.  It creates an unique id for each athlete.  The athlete id follows the format used by USA Swimming. Bithdate First; 3 letters of first name; first 4 leters of last name. (mm/dd/yyseawarl)
}

