% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_swims.R
\name{individual_swims}
\alias{individual_swims}
\title{Retrive Report From USA Swimming's NCAA DI Top Times Report}
\usage{
individual_swims(Conference, start_date = "2015-08-01",
  end_date = "2015-05-31", LCM = FALSE, SCM = FALSE, SCY = TRUE, top,
  downloadPath = getwd(), read = TRUE)
}
\arguments{
\item{Conference}{a character string.  The name of a Division 1 conference.}

\item{start_date}{a character string.   Defines the begining of the period for the times report. It is entered as a character string in the format 'YYYY-MM-DD'.}

\item{end_date}{a character string.   Defines the end of the period for the times report. It is entered as a character string in the format 'YYYY-MM-DD'.}

\item{LCM}{logical.  Optional. Should long course meter events be included in the report?  Defaults to FALSE.}

\item{SCM}{logical.  Optional. Should short course meter events be included in the report?  Defaults to FALSE.}

\item{SCY}{logical.  Optional. Should short course yard events be included in the report?  Defaults to TRUE.}

\item{top}{integer. How many swimmers should be returned in each event? Maximum allowed is 500.}

\item{downloadPath}{character sting}

\item{read}{logical}
}
\description{
\code{individual_swims} fills out the report search form to retrun a top time report.
}

