% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_fucntions.R
\name{insert_athlete}
\alias{insert_athlete}
\title{Insert Athlete Data into database.
An Internal function.}
\usage{
insert_athlete(con, data)
}
\arguments{
\item{con}{a}

\item{data}{}
}
\description{
Insert Athlete Data into database.
An Internal function.
}
\keyword{internal}
