% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_fucntions.R
\name{insert_conference}
\alias{insert_conference}
\title{Insert Team Data into database.
An Internal function.}
\usage{
insert_conference(con, conference_name)
}
\arguments{
\item{con}{a}

\item{conference_name}{}
}
\description{
Insert Team Data into database.
An Internal function.
}
\keyword{internal}
