% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_fucntions.R
\name{insert_event}
\alias{insert_event}
\title{Insert Event Data into database.
An Internal function.}
\usage{
insert_event(con)
}
\arguments{
\item{con}{a database connection of class}

\item{data}{}
}
\description{
Insert Event Data into database.
An Internal function.
}
\keyword{internal}
