% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_fucntions.R
\name{insert_meet}
\alias{insert_meet}
\title{Insert Meet Data into database.
An Internal function.}
\usage{
insert_meet(con, data)
}
\arguments{
\item{con}{a}

\item{data}{}
}
\description{
Insert Meet Data into database.
An Internal function.
}
\keyword{internal}
