% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_fucntions.R
\name{insert_result}
\alias{insert_result}
\title{Insert Result Data into database.
An Internal function.}
\usage{
insert_result(con, data)
}
\arguments{
\item{con}{a}

\item{data}{}
}
\description{
Insert Result Data into database.
An Internal function.
}
\keyword{internal}

