% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_fucntions.R
\name{insert_team}
\alias{insert_team}
\title{Insert Team Data into database.
An Internal function.}
\usage{
insert_team(con, data, conference_name)
}
\arguments{
\item{con}{a}

\item{data}{}
}
\description{
Insert Team Data into database.
An Internal function.
}
\keyword{internal}

