% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{interpolate_exponential}
\alias{interpolate_exponential}
\title{Exponentially interpolate around two data points}
\usage{
interpolate_exponential(p1, p2, t1, t2, t0)
}
\arguments{
\item{p1}{Value in period 1}

\item{p2}{Value in period 2}

\item{t1}{Time step for period 1 (i.e., 2010)}

\item{t2}{Time step for period 2 (i.e., 2040)}

\item{t0}{The timepoint for interpolation (or extrapolation)}
}
\description{
Exponentially interpolate around two data points
}
\details{
This function first calculates the implied exponential growth rate
  between two points, and then predicts where the function would be in the
  intermediate year.
}
\examples{
interpolate_exponential(10, 12, 2010, 2040, 2020)
interpolate_exponential(10, 12, 2010, 2040, 2008)
}
\seealso{
calc_exprate
}
