% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logsums.R
\name{map_logsums}
\alias{map_logsums}
\title{Map destination choice logsums from a single scenario}
\usage{
map_logsums(
  db,
  scope = NULL,
  purposes = NULL,
  ggmap = FALSE,
  show_year = 2010,
  index_year = 2000
)
}
\arguments{
\item{db}{The scenario sqlite database.}

\item{scope}{a dplyr::filtering criteria to limit the scope of the dataframe}

\item{purposes}{a vector of trip purposes to include in the average logsum.}

\item{ggmap}{If TRUE, then include a ggmap background.}

\item{show_year}{The year in which to plot the logsums.}

\item{index_year}{[Optional] index year that should be used as the starting year for data or plots;}
}
\description{
Map destination choice logsums from a single scenario
}
