% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mode_types}
\alias{mode_types}
\title{Consolidated transport modes}
\format{A data frame with 9 observations.
\describe{
  \item{mode}{The mode in the TRIPMATRIX} \item{consolidated_mode}{One of
  five modes: auto, transit, school, truck, and non-motorized.}
}}
\usage{
mode_types
}
\description{
A dataset mapping modes in the trip matrix to consolidated values for plotting.
}
\keyword{datasets}

