% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employment.R
\name{multiple_employment}
\alias{multiple_employment}
\title{Compare Employment by sector across multiple scenarios.}
\usage{
multiple_employment(dbset, db_names, ...)
}
\arguments{
\item{dbset}{A list of connections to SWIM databases.}

\item{db_names}{A character vector naming the scenarios.}
}
\value{
a ggplot2 object.
}
\description{
Compare Employment by sector across multiple scenarios.
}

