% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdp.R
\name{multiple_gdp}
\alias{multiple_gdp}
\title{Compare GDP across multiple scenarios.}
\usage{
multiple_gdp(
  dbset,
  db_names,
  facet_var = c("MPO", "COUNTY", "STATE"),
  facet_levels = NULL,
  index_year = 2000
)
}
\arguments{
\item{dbset}{A list of connections to SWIM databases.}

\item{db_names}{A character vector naming the scenarios.}

\item{facet_var}{Field to facet by.}

\item{facet_levels}{A character vector of the facet variable specifiying
which levels to include.}

\item{index_year}{[Optional] index year that should be used as the starting year for data or plots;}
}
\value{
a ggplot2 object.
}
\description{
Compare GDP across multiple scenarios.
}
