% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trips.R
\name{multiple_trips}
\alias{multiple_trips}
\title{Plot trips in multiple scenarios.}
\usage{
multiple_trips(dbset, db_names, index_year = 2000, ...)
}
\arguments{
\item{dbset}{A list of connections to SWIM databases.}

\item{db_names}{A character vector naming the scenarios.}

\item{...}{
  Arguments passed on to \code{\link[=extract_trips]{extract_trips}}
  \describe{
    \item{\code{db}}{the scenario database.}
    \item{\code{facet_var}}{The region to dplyr::summarize by.}
    \item{\code{facet_levels}}{Regions to include in summary.}
    \item{\code{color_levels}}{Modes to include in summary. Defaults to all modes other
than \code{school}. See consolidated modes in \link{mode_types}.}
    \item{\code{index}}{whether to index mode split off the base year.}
    \item{\code{index_year}}{[Optional] index year that should be used as the starting year for data or plots;}
  }}
}
\value{
A ggplot2 object
}
\description{
Plot trips in multiple scenarios.
}
