% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmt.R
\name{multiple_vmt}
\alias{multiple_vmt}
\title{Compare VMT between multiple scenarios}
\usage{
multiple_vmt(dbset, db_names, facet_var = c("MPO", "COUNTY", "STATE"),
  facet_levels = NULL)
}
\arguments{
\item{dbset}{A list of connections to SWIM databases.}

\item{db_names}{A character vector naming the scenarios.}

\item{facet_var}{The region to summarize by.}

\item{facet_levels}{Regions to include in summary.}
}
\value{
A ggplot2 object showing VHT by facility type in each facet level
  over time.
}
\description{
Compare VMT between multiple scenarios
}

