% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{pct_diff}
\alias{pct_diff}
\title{Percent difference between two conditions}
\usage{
pct_diff(x, y)
}
\arguments{
\item{x}{base value}

\item{y}{alternate value}
}
\description{
Percent difference between two conditions
}

