% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_validation.R
\name{plot_countcomparison}
\alias{plot_countcomparison}
\title{Plot base-year traffic validation.}
\usage{
plot_countcomparison(db, year = c(2010, 2013))
}
\arguments{
\item{db}{The scenario database.}

\item{facet_var}{Variable to display in facets}
}
\value{
A ggplot2 object.
}
\description{
Create a plot of the base scenario link volumes compared with observed AADT.
}

