% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employment.R
\name{plot_employment}
\alias{plot_employment}
\title{Plot Employment by Sector}
\usage{
plot_employment(db, ...)
}
\value{
A ggplot2 object showing the employment by type and and year.
}
\description{
This function plots the employment by type in an area over time between two
scenarios.
}

