% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floorspace.R
\name{plot_floorspace}
\alias{plot_floorspace}
\title{Plot floorspace statistics over time}
\usage{
plot_floorspace(..., price = FALSE)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=extract_rents]{extract_rents}}
  \describe{
    \item{\code{db}}{The scenario database.}
    \item{\code{facet_var}}{The variable in the zone table to facet by. Defaults to MPO}
    \item{\code{facet_levels}}{The levels of the facet variable to keep. Defaults to all
levels other than external stations.}
    \item{\code{type_levels}}{The types of employment to show in the plot.}
    \item{\code{index}}{Should the function extract indexed or absolute values?}
    \item{\code{index_year}}{[Optional] index year that should be used as the starting year for data or plots;}
  }}

\item{price}{Print price instead of floorspace, defaults to \code{FALSE}.}
}
\value{
A ggplot2 object showing the floorspace by type and and year.
}
\description{
This function plots the constructed floorspace by type over time facetted by a
variable chosen from the zone attributes table.
}
