% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{plot_history}
\alias{plot_history}
\title{Plot population with controls and historical data.}
\usage{
plot_history(db, counties = NULL)
}
\arguments{
\item{db}{Swim database.}

\item{counties}{If not null (default), a character vector of counties to plot.}
}
\value{
a ggplot2 object
}
\description{
Plot population with controls and historical data.
}

