% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/laborforce.R
\name{plot_lfpr_volatility}
\alias{plot_lfpr_volatility}
\title{Plot LFPR Volatility in a Region}
\usage{
plot_lfpr_volatility(db, level = c("BZONE", "COUNTY", "MPO", "ALDREGION",
  "STATE"), scope = NULL, ggmap = FALSE)
}
\arguments{
\item{db}{Scenario database.}

\item{level}{Level at which to calculate volatility over time. Smaller
levels, such as BZONE will show higher volatility.}

\item{scope}{a filtering criteria to limit the scope of the}

\item{ggmap}{if TRUE, return a ggmap background}
}
\value{
a ggplot2 object.
}
\description{
This function calculates the volatility of the labor force participation rate
at an arbitrary geographic level. Volatility is defined as the standard
deviation of the percent growth rate of LFPR between consecutive periods.
}

