% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelwide.R
\name{plot_modelwide}
\alias{plot_modelwide}
\title{Plot model wide parameters.}
\usage{
plot_modelwide(db, filters)
}
\arguments{
\item{db}{Scenario database.}

\item{filters}{list object containing character strings on which to filter
parameters.}
}
\description{
Plot model wide parameters.
}

