% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vmt.R
\name{plot_pct_cong}
\alias{plot_pct_cong}
\title{Plot percent congested links over time.}
\usage{
plot_pct_cong(db, color_var = "MPO", color_levels = NULL,
  congested_voc = 0.9)
}
\arguments{
\item{db}{The scenario database.}

\item{color_var}{The variable to use in coloring the data.}

\item{color_levels}{A character vector giving the levels of the coloring
  variable to use. Levels not called are dropped from the plot; default is
  \code{NULL}, meaning print all level.}
}
\value{
A ggplot2 figure object.
}
\description{
This function plots the percent of congested links, colored and
faceted by arbitrary variables in the link table.
}

