% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/population.R
\name{plot_rates}
\alias{plot_rates}
\title{Plot annualized growth rates}
\usage{
plot_rates(db, counties)
}
\arguments{
\item{db}{the Swim database.}

\item{counties}{A character vector of counties to display.}
}
\description{
This function returns a box plot comparing the annualized population growth
rates in a set of counties. This can help identify problems that the model
may be having relative to historical trends.
}

