% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_sevar}
\alias{plot_sevar}
\title{Make a plot of population and employment}
\usage{
plot_sevar(db, color_var = c("MPO", "COUNTY"), color_levels = NULL,
  controls = TRUE)
}
\arguments{
\item{db}{The scenario sqlite database.}

\item{color_var}{Field to color by: either "MPO" or "COUNTY".}

\item{color_levels}{A character vector of the color variable specifiying
which levels to include.}

\item{controls}{Plot against the control totals. Defaults to TRUE, cannot
  currently run with FALSE.}
}
\value{
A \code{ggplot2} plot object showing the modeled change in employment
  and population over time.
}
\description{
Make a plot of population and employment
}

