% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlfd.R
\name{plot_tlfd}
\alias{plot_tlfd}
\title{Plot a trip length frequency distribution}
\usage{
plot_tlfd(
  db,
  region_var = c("MPO", "COUNTY", "STATE"),
  regions = NULL,
  cumulative = FALSE,
  years = NULL,
  index_year = 2000
)
}
\arguments{
\item{db}{The scenario database.}

\item{region_var}{The region to aggregate to.}

\item{regions}{The regions to return.  If \code{NULL}, returns all.}

\item{cumulative}{Plot cumulative instead of probability density distribution.}

\item{years}{The years to show in the plot. If null, will show the first
and last years.}

\item{index_year}{[Optional] index year that should be used as the starting year for data or plots;}
}
\value{
A ggplot2 object.
}
\description{
Plot a trip length frequency distribution
}
