% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_validation.R
\name{plot_traffic_count}
\alias{plot_traffic_count}
\title{Plot historical and projected traffic counts}
\usage{
plot_traffic_count(db, atr = c(1001, 1011, 1012))
}
\arguments{
\item{db}{The scenario database.}

\item{atr}{A numeric vector indicating the traffic recorders to plot track.}
}
\value{
A ggplot2 object
}
\description{
Plot historical and projected traffic counts
}

