% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trips.R
\name{plot_trips}
\alias{plot_trips}
\title{Plot Trips by MPO}
\usage{
plot_trips(db, facet_var = "MPO", facet_levels = NULL, share = TRUE)
}
\arguments{
\item{db}{the scenario database.}

\item{facet_var}{Defaults to MPO}

\item{facet_levels}{defaults to all}

\item{share}{Plot mode split.}
}
\description{
Plot Trips by MPO
}

