% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_validation.R
\name{plot_validation}
\alias{plot_validation}
\title{Plot base-year traffic validation.}
\usage{
plot_validation(db)
}
\arguments{
\item{db}{The scenario database.}
}
\value{
A ggplot2 object.
}
\description{
Create a plot of the base scenario link volumes compared with observed AADT.
}

