% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vmt.R
\name{plot_vht}
\alias{plot_vht}
\title{Plot VHT over time.}
\usage{
plot_vht(db, color_var = "MPO", color_levels = NULL)
}
\arguments{
\item{db}{The scenario database.}

\item{color_var}{The variable to use in coloring the data.}

\item{color_levels}{A character vector giving the levels of the coloring
  variable to use. Levels not called are dropped from the plot; default is
  \code{NULL}, meaning print all level.}
}
\value{
A ggplot2 figure object.
}
\description{
This function plots the sum of the daily vehicle travel time, colored and
faceted by arbitrary variables in the link or zone table.
}

