% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laborforce.R
\name{plot_wapr_volatility}
\alias{plot_wapr_volatility}
\title{Plot WAPR Volatility in a Region}
\usage{
plot_wapr_volatility(
  db,
  level = c("BZONE", "COUNTY", "MPO", "ALDREGION", "STATE"),
  scope = NULL,
  index_year = 2000
)
}
\arguments{
\item{db}{Scenario database.}

\item{level}{Level at which to calculate volatility over time. Smaller
levels, such as BZONE will show higher volatility.}

\item{scope}{a dplyr::filtering criteria to limit the scope of the dataframe}

\item{index_year}{[Optional] index year that should be used as the starting year for data or plots;}
}
\value{
a ggplot2 object.
}
\description{
This function calculates the volatility of worker to adult population ratio
at an arbitrary geographic level. Volatility is defined as the standard
deviation of the percent growth rate of wapr between consecutive periods.
}
