% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{pop_rate}
\alias{pop_rate}
\title{Determine annualized population growth rates.}
\usage{
pop_rate(db, counties = NULL)
}
\arguments{
\item{db}{Swim database.}

\item{counties}{If not null (default), a character vector of counties to plot.}
}
\description{
This function calculates the annualized growth rate between Census years and
between SWIM population synthesizer runs.
}
