% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firefox profile.R
\name{profile}
\alias{profile}
\title{Build Custom Firefox profile}
\usage{
profile(downloadPath)
}
\arguments{
\item{downlaodPath}{a character string with a file path.}
}
\description{
Profile is an internal function used by \code{\link{individual_swims}}
}
\details{
Profile is an internal function used by \code{\link{individual_swims}}.  It prepares a Firefox profile so that the Top Times report generated by \code{\link{individual_swims}} can be downloaded as a \emph{.csv} without any human interaction.  It is a wraper around \code{\link[RSelenium]{makeFirefoxProfile}}.
}
\seealso{
\code{\link{individual_swims}} \code{\link[RSelenium]{makeFirefoxProfile}}
}
\keyword{internal}
