% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_prep.R
\name{query_prep}
\alias{query_prep}
\title{Prepare SQL Query}
\usage{
query_prep(conference = NULL, team_name = NULL, team_code = NULL,
  gender = "Both", athlete = NULL, event = "All")
}
\arguments{
\item{conference}{a character vector with the name of an NCAA swimming conference.  If not provided the results include all conferences.}

\item{team_name}{A character vector with the name of an NCAA swimming conference.  If not provided the results include all teams}

\item{gender}{a character vector.  Indicates the output should include men, women or both.  Defaults to "Both".  Other options are "M" for men and "F" for women.}

\item{athlete}{a character vector with the name or athlete id of a swimmier.  If not provided the results include all swimmers.  The athlete id follows the USA Swimming standard.}

\item{event}{a character vector with the name of swimming envent.  If not provided the results include all events.}

\item{team_name}{a character vector with the team code of an NCAA swimming team  If not provided the results include all teams}
}
\description{
The function prepares a SQL query based on the parameters provided.
}
\keyword{internal}

