% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_counts}
\alias{ref_counts}
\title{Reference counts}
\format{A data frame with 167 observations.
\describe{
  \item{site}{The site location of the count.}
  \item{AADT}{Observed traffic flow.}
  \item{year}{The year of the count.}
  \item{FROMNODENO}{The from node of the link in the SWIM highway network.}
  \item{TONODENO}{The from node of the link in the SWIM highway network.}
}}
\source{
Calculated by Rick Donnelly from data supplied by ODOT.
}
\usage{
ref_counts
}
\description{
A dataset containing the reference highway counts for comparing modeled
traffic volumes in the calibration year.
}
\keyword{datasets}

