% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regions}
\alias{regions}
\title{Regions table}
\format{An object of class \code{data.frame} with 75 rows and 2 columns.}
\usage{
regions
}
\description{
A table relating ODOT regions to counties.
}
\keyword{datasets}
