% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lookup.R
\name{remove_field}
\alias{remove_field}
\title{Remove a column from a database table}
\usage{
remove_field(db, table, column)
}
\arguments{
\item{db}{A scenario database.}

\item{table}{The name of the table in the database.}

\item{column}{The name of the column to drop.}
}
\description{
This might be useful if you make a mistake with \link{add_lookup}.
As a warning, this is a destructive operation. Use with care.
}

