% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_ncaa_qualifiers.R
\name{report_ncaa_qualifiers}
\alias{report_ncaa_qualifiers}
\title{Determine the number NCAA Swimming Championship qualifiers}
\usage{
report_ncaa_qualifiers(con, conference = "All", team_name = "All",
  gender = "Both", athlete = NULL, event = "All")
}
\arguments{
\item{con}{a database connection.}

\item{conference}{a character vector with the name of an NCAA swimming conference.  If not provided the output include all conferences.}

\item{team_name}{A character vector with the name of an NCAA swimming conference.  If not provided the output include all teams.}

\item{gender}{a character vector.  Indicates the output should include men, women or both.  Defaults to "Both".  Other options are "M" for men and "F" for women.}

\item{athlete}{a character vector with the name or athlete id of a swimmier.  If not provided the results include all swimmers.  The athlete id follows the USA Swimming standard.}

\item{event}{a character vector with the name of swimming envent.  If not provided the results include all events.}
}
\value{
a data frame providing a count of ncaa qualifiers aggregated 
according to the arguments provided.
}
\description{
Determine the number NCAA Swimming Championship qualifiers
}
\details{
\code{report_top_times} queries a SQL database to produce a ranking of swimmers in the specified events.  The function is flexiable and can rank swimmers across all of Division 1, within a specific conference or within a team.  It can also produce an individual swimmer's best result in each event they have competed in.

When producing the count the function aggregates along the specified
arguments.  If left at the defult the argument is not included in the
aggregate grouping.
}

