% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWIMR_prop.R
\name{selectInfo}
\alias{selectInfo}
\title{Select Informative Genes used in the deconvolution}
\usage{
selectInfo(Basis, sc_eset, commonGene, ct.select = NULL, ct.varname)
}
\arguments{
\item{Basis}{A reference basis matrix where columns are cell types
and rows are genes. Values represent gene expression.}

\item{sc_eset}{A SingleCellExperiment S4 object containing scRNA-seq
data and metadata.}

\item{commonGene}{A character vector of genes common to both
the single-cell and spatial transcriptomics datasets.}

\item{ct.select}{A character vector of cell type names to be
used for deconvolution. If NULL, all cell types are used.}

\item{ct.varname}{A character string specifying the column name
in `colData(sc_eset)` that contains cell type annotations.}
}
\value{
A character vector of selected informative genes.
}
\description{
This function selects a set of informative genes for deconvolution
based on a combination of differential expression and gene dispersion
within cell types.
}
