% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenium_start.R
\name{selenium_start}
\alias{selenium_start}
\title{Start Selenium Server From R}
\usage{
selenium_start()
}
\description{
Start Selenium Server From R
}
\details{
To Do:
* [ ] Set up function to work with chosen directory.  
* [ ] Document Function.
}

