% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim_display.R
\name{skim_leaflet}
\alias{skim_leaflet}
\title{Plot a skim chloropleth}
\usage{
skim_leaflet(zone, skim, dir1, dir2 = NULL, year1, year2 = NULL,
  from = TRUE)
}
\arguments{
\item{zone}{The focus zone.}

\item{skim}{The name of the skim file in the \code{dir1/outputs/t*/}
directory.}

\item{dir1}{The base or reference directory.}

\item{dir2}{The alternative or analysis scenario. Default \code{NULL}
scenario.}

\item{year1}{The base or reference year to show, e.g., \code{t20}.}

\item{year2}{The comparison or analysis year. Default \code{NULL}.}

\item{from}{If \code{TRUE}, gets values from \code{zone} to other zones.}
}
\description{
Plot a skim chloropleth
}
\details{
If \code{dir2} and \code{year2} are both \code{NULL}, will show an
  absolute measure of travel time. If one or the other is given, will show
  a percent change from \code{dir1, year1}.
}
