% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim_display.R
\name{skim_popup}
\alias{skim_popup}
\title{Build skim popup tag for leaflet}
\usage{
skim_popup(shp, zone)
}
\arguments{
\item{zone}{The focus zone.}

\item{var}{Variable to create popup tag for (e.g., \code{"pop"})}
}
\description{
Build skim popup tag for leaflet
}

