% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_parse.R
\name{start_dateParse}
\alias{start_dateParse}
\title{Parse & Format Dates For {RSelenium}'s \code{remoteDriver}}
\usage{
start_dateParse(start_date)
}
\arguments{
\item{start_date}{}
}
\value{
a list.  Each character from the start date is an element of the list.
}
\description{
\code{start_dateParse} is an internal function used in \code{individual_swims}.
}
\details{
\code{start_dateParse} is an internal function used in \code{individual_swims}.  It prepares the \emph{start_date} argument in \code{individual_swims} so that \strong{RSelenium} can pass it to the date fields on \href{http://www.usaswimming.org/DesktopDefault.aspx?TabId=1971&Alias=Rainbow&Lang=en}{USA Swimming's Top Times Report} form.  It converts each character in \emph{start_date} into an element of a list.
}
\seealso{
\code{\link[RSelenium]{remoteDriver}}
}

