% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenium_start.R
\name{start_selenium}
\alias{start_selenium}
\title{Start Selenium Standalone Server From R}
\usage{
start_selenium(dir = NULL)
}
\arguments{
\item{dir}{A directory where the Selenium Standalone Server file is located}
}
\value{
Returns a list of two functions.  \code{getPID} provides the process id for the running Selenium server.  \code{stop} function utilizes the process id to stop the Selenium server.
}
\description{
Start Selenium Standalone Server From R
}
\details{
The function is based \code{\link[RSelenium]{startServer}}, which is now depricated.  In particular returned value is based on the source code of the former function.

While the function replicates some of the features of \code{\link[RSelenium]{startServer}}, it is not as robust.  Currently the function is designed and has only been tested on macOS.
}
