% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swim_db_data_load.R
\name{swim_db_data_load}
\alias{swim_db_data_load}
\title{Load Data Frame Into Swim Database}
\usage{
swim_db_data_load(con, data, conference_name)
}
\description{
Load Data Frame Into Swim Database
}

