% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swimdb_tables.R
\name{swim_db_tables}
\alias{swim_db_tables}
\title{Create tables to store data in swimming database.}
\usage{
swim_db_tables(con)
}
\arguments{
\item{con}{A character string providing the name of a SQLiteConnection to the desired database.}
}
\value{
The function prints a message to the terminal indicating if the table was correctly created.
}
\description{
Create tables to store data in swimming database.
}
\details{
\code{swim_db_tables} is used to create tables in a SQLite database.  These tables are designed to store results from USA Swimmings Top Times Report.  The function utilizes \code{\link[DBI]{DBI}} create the connection to the database, creation of the tables and insertion of the data.

If the database provided in /emph{db_name} does not exist the fuction will create the SQLite database.
}

