% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{class}
\name{SWIMR-class}
\alias{SWIMR-class}
\title{Each SWIMR object has a number of slots which store information. Key slots to access are listed below.}
\description{
Each SWIMR object has a number of slots which store information. Key slots to access are listed below.
}
\section{Slots}{

\describe{
\item{\code{sc_eset_list}}{A list of filtered scRNA-seq datasets, together with
associated metadata, stored as SingleCellExperiment objects.}

\item{\code{spatial_count}}{The filtered spatial transcriptomics count data.}

\item{\code{spatial_location}}{A data frame specifying spatial coordinates for each location.}

\item{\code{Weight_est}}{A matrix of estimated weights, where rows correspond to spatial
locations and columns to scRNA-seq references.}

\item{\code{Proportion_est}}{A matrix of estimated cell-type proportions from SWIMR,
with rows representing spatial locations and columns representing cell types.}

\item{\code{Proportion_single}}{A list of estimated cell-type proportions from SWIMR from each reference.}

\item{\code{info_parameters}}{A list of parameters used during model fitting.}
}}

