\name{SwimR-package}
\alias{SwimR-package}

\docType{package}
\title{
 SwimR: A Suite of Analytical Tools for Quantification of C. elegans Swimming Behavior
}

\description{
 SwimR is an R-based suite that calculates, analyses, and plots the frequency of 
 C. elegans swimming behavior over time.  It places a particular emphasis on 
 identifying paralysis and quantifying the kinetic elements of paralysis 
 during swimming. Data is input to SwipR from a custom built program that 
 fits a 5 point morphometric spine to videos of single worms swimming 
 in a buffer called Worm Tracker.
}

\details{
    \tabular{ll}{
        Package: \tab SwimR\cr
        Type: \tab Package\cr
        Version: \tab 0.99.1\cr
        Date: \tab 2013-08-17\cr
        License: \tab LGPL (>= 2)\cr
        LazyLoad: \tab yes\cr
    }
}

\author{
    Jing Wang, Andrew Hardaway, Bing Zhang
    Maintainer: Randy Blakely <Randy.Blakely@Vanderbilt.edu>
}

\references{
 An Open-Source Analytical Platform for Analysis of C. elegans Swimming 
 Induced Paralysis (under reivew).
}

\seealso{
    \code{\link{SwimR}}
}

\keyword{ package }