% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWIMR_prop.R
\name{SWIMR_deconvolution}
\alias{SWIMR_deconvolution}
\title{Run multi-reference deconvolution on a SWIMR object}
\usage{
SWIMR_deconvolution(obj)
}
\arguments{
\item{obj}{A \code{SWIMR} object containing:
\itemize{
  \item \code{sc_eset_list}: list of filtered scRNA-seq references
  \item \code{spatial_count}: filtered spatial count matrix
  \item \code{spatial_location}: matched coordinates data frame
  \item \code{info_parameters}: list of key parameters used for construction
}}
}
\value{
A \code{SWIMR} object with updated slots:
\itemize{
  \item \code{Weight_est}: reference weights per spatial location
  \item \code{Proportion_est}: fused cell type proportions per spatial location
  \item \code{algorithm_matrix}: intermediates from CARD
}
}
\description{
Run multi-reference deconvolution on a SWIMR object
}
