% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_convert.R
\name{time_convert}
\alias{time_convert}
\title{Convert time to duration for calculations.}
\usage{
time_convert(data)
}
\arguments{
\item{data}{The data frame provided in \code{\link{data_clean}}}
}
\description{
An internal function used in \code{\link{data_clean}}
}
\details{
\code{time_convert} is an internal function used in \code{\link{data_clean}} to convert the athletes time from a character type to a variable type that can be used in calculations.  It utilizes \code{\link[lubridate]{lubridate}} to perform the conversion.
}

