% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_troubleshoot.R
\name{troubleshoot_aa}
\alias{troubleshoot_aa}
\title{AA Troubleshooter}
\usage{
troubleshoot_aa(db)
}
\arguments{
\item{db}{A scenario database.}
}
\value{
A [DT](`datatable`) object.
}
\description{
This function prints aggregate statistics from the AA buy/sell matrix. This
can help an analyst determine which commodities (if any) may cause problems.
}

