% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{yearly_summary}
\alias{yearly_summary}
\title{Make a yearly summary table of a land use variable}
\usage{
yearly_summary(df, group, var)
}
\arguments{
\item{df}{A \code{data_frame} of zones, such as is stored in the sqlite
database.}

\item{group}{The variable to use for grouping the zones (for instance,
"County").}

\item{var}{The variable to sum (for instance, "POPULATION")}
}
\value{
A \code{data_frame} with the grouping variable in each row, and the
value of the variable at each time stage going across the columns.
}
\description{
Make a yearly summary table of a land use variable
}

