% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone_leaflet.R
\name{zone_leaflet}
\alias{zone_leaflet}
\title{Construct a base leaflet zone plot}
\usage{
zone_leaflet(shp)
}
\arguments{
\item{shp}{An object of class \code{SpatialPolygonsDataFrame} with the
plotting fields already appended.}
}
\description{
This function creates a standard leaflet base image that we can add
information to.
}

