% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zones}
\alias{zones}
\title{Zone table}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 701344 rows and 15 columns.}
\usage{
zones
}
\description{
A fortified shapefile containing the zone geometries
}
\keyword{datasets}

