% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zones_shp}
\alias{zones_shp}
\title{Zone Shapefile}
\format{An object of class \code{SpatialPolygonsDataFrame} with 2986 rows and 9 columns.}
\usage{
zones_shp
}
\description{
A SpatialPolygonsDataFrame of the zone system for plotting in leaflet.
}
\keyword{datasets}

