% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/employment.R
\name{compare_employment}
\alias{compare_employment}
\title{Compare Employment between two scenarios}
\usage{
compare_employment(db1, db2, facet_var = c("MPO", "COUNTY", "STATE"),
  facet_levels = NULL, type_levels = NULL)
}
\arguments{
\item{db1}{The reference scenario database.}

\item{db2}{The current scenario database.}

\item{facet_var}{The variable in the zone table to facet by. Defaults to MPO}

\item{facet_levels}{The levels of the facet variable to keep. Defaults to all
levels other than external stations.}

\item{type_levels}{The types of employment to show in the plot.}
}
\value{
A ggplot2 object showing the employment by type and and year.
}
\description{
This function plots the value of employment sold by a zone. This is a proxy
for the number of workers in each sector, which is not immediately available
from the SWIM database.
}

