% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/floorspace.R
\name{compare_floorspace}
\alias{compare_floorspace}
\title{Compare floorspace over time}
\usage{
compare_floorspace(db1, db2, facet_var = c("MPO", "COUNTY", "STATE"),
  facet_levels = NULL, type_levels = NULL)
}
\arguments{
\item{db1}{The swim database for the "Reference" scenario.}

\item{db2}{The swim database for the "Current" scenario.}

\item{facet_var}{The variable in the zone table to facet by. Defaults to MPO}

\item{facet_levels}{The levels of the facet variable to keep. Defaults to all
levels other than external stations.}

\item{type_levels}{The types of floorspace to show in the plot.}
}
\description{
Compare floorspace over time
}

