% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlfd.R
\name{compare_tlfd}
\alias{compare_tlfd}
\title{Compare Trip Length Frequency Distributions}
\usage{
compare_tlfd(
  db1,
  db2,
  region_var = c("MPO", "COUNTY", "STATE"),
  regions = NULL,
  cumulative = FALSE,
  years = NULL,
  index_year = 2000
)
}
\arguments{
\item{db1}{The swim database for the "Reference" scenario.}

\item{db2}{The swim database for the "Current" scenario.}

\item{region_var}{The region to aggregate to.}

\item{regions}{The regions to return.  If \code{NULL}, returns all.}

\item{cumulative}{Plot cumulative instead of probability density distribution.}

\item{years}{The years to show in the plot. If null, will show the first
and last years.}

\item{index_year}{[Optional] index year that should be used as the starting year for data or plots;}
}
\value{
A ggplot2 object.
}
\description{
Compare Trip Length Frequency Distributions
}
