% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trips.R
\name{compare_trips}
\alias{compare_trips}
\title{Compare Trips}
\usage{
compare_trips(db1, db2, facet_var = "MPO", facet_levels = NULL)
}
\arguments{
\item{db1}{The swim database for the "Reference" scenario.}

\item{db2}{The swim database for the "Current" scenario.}

\item{facet_var}{Field to facet by: either "MPO" or "COUNTY".}

\item{facet_levels}{A character vector of the facet variable specifiying
  which levels to include.}
}
\description{
Compare Trips
}

