% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone_leaflet.R
\name{diff_leaflet}
\alias{diff_leaflet}
\title{Leaflet plot of difference between two scenarios.}
\usage{
diff_leaflet(db1, db2, year, variable = c("Population", "Employment", "HH"),
  scen_names = c("Reference", "Current"))
}
\arguments{
\item{db1}{Reference scenario VIZ database connection.}

\item{db2}{Current scenario VIZ database connection.}

\item{year}{Year in which to compare the two scenarios.}

\item{variable}{Which variable to show in difference plot; currently
\code{variable = c("Population", "Employment", "HH")}}

\item{scen_names}{Names of the scenarios in the comparison. Defaults to
"Reference", "Current".}
}
\description{
Leaflet plot of difference between two scenarios.
}

