% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/employment.R
\name{extract_employment}
\alias{extract_employment}
\title{Extract employment statistics from scenario}
\usage{
extract_employment(db, facet_var = c("MPO", "COUNTY", "STATE"),
  facet_levels = NULL, type_levels = NULL)
}
\arguments{
\item{db}{The scenario database.}

\item{facet_var}{The variable in the zone table to facet by. Defaults to MPO}

\item{facet_levels}{The levels of the facet variable to keep. Defaults to all
levels other than external stations.}

\item{type_levels}{The types of employment to show in the plot.}
}
\description{
This is an internal function to pull the employment data from a scenario and
return it to either the single scenario or scenario comparison functions.
}

