% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/economics.R
\name{extract_floorspace}
\alias{extract_floorspace}
\title{Extract floorspace statistics from scenario}
\usage{
extract_floorspace(db, facet_var = c("MPO", "COUNTY", "STATE"),
  facet_levels = NULL, type_levels = NULL)
}
\description{
This is an internal function to pull the floorspace data from a scenario and
return it to either the single scenario or scenario comparison functions.
}

