% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdp.R
\name{extract_gdp}
\alias{extract_gdp}
\title{Extract GDP of Labor}
\usage{
extract_gdp(
  db,
  facet_var = c("MPO", "COUNTY", "STATE"),
  facet_levels = NULL,
  color_levels = NULL,
  index_year = 2000
)
}
\arguments{
\item{db}{The scenario sqlite database.}

\item{facet_var}{Field to facet by: "MPO", "COUNTY", or "STATE".}

\item{facet_levels}{A character vector of the variable specifiying
which levels to include.}

\item{color_levels}{A character vector of the industry sectors to include.
Defaults to all.}

\item{index_year}{[Optional] index year that should be used as the starting year for data or plots;}
}
\value{
A \code{data_frame} with the participation rate in each facet region
  in each transport model year.
}
\description{
This function returns the total dollar value of all the labor sold by a
dplyr::collect(n=Inf)ion of zones over time.
}
