% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlfd.R
\name{extract_tlfd}
\alias{extract_tlfd}
\title{Extract trip length frequency distribution data}
\usage{
extract_tlfd(db, region_var = c("MPO", "COUNTY", "STATE"), regions = NULL,
  bin_width = 1, max_bin = 100)
}
\arguments{
\item{db}{The scenario database.}

\item{region_var}{The region to aggregate to.}

\item{regions}{The regions to return.  If \code{NULL}, returns all.}

\item{bin_width}{The width of bins to use in the trip length frequency
distribution. Defaults to 1.}

\item{max_bin}{The maximum bin size to return.}
}
\value{
A data frame with the trip length frequency distribution by year
  and region.
}
\description{
Extract trip length frequency distribution data
}

