% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trips.R
\name{extract_trips}
\alias{extract_trips}
\title{Extract trips from scenario}
\usage{
extract_trips(db, facet_var = "MPO", facet_levels = NULL)
}
\arguments{
\item{db}{the scenario database.}

\item{facet_var}{Defaults to MPO}

\item{facet_levels}{defaults to all}
}
\description{
Extract trips from scenario
}

