% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trips.R
\name{extract_trips}
\alias{extract_trips}
\title{Extract trips from scenario}
\usage{
extract_trips(
  db,
  facet_var = NULL,
  facet_levels = NULL,
  color_levels = c("auto", "transit", "non-motorized", "truck"),
  index = FALSE,
  index_year = 2000
)
}
\arguments{
\item{db}{the scenario database.}

\item{facet_var}{The region to dplyr::summarize by.}

\item{facet_levels}{Regions to include in summary.}

\item{color_levels}{Modes to include in summary. Defaults to all modes other
than \code{school}. See consolidated modes in \link{mode_types}.}

\item{index}{whether to index mode split off the base year.}

\item{index_year}{[Optional] index year that should be used as the starting year for data or plots;}
}
\value{
A data frame with total trips originating in the region by mode and
  year.
}
\description{
This function looks at the trip matrix for a scenario and dplyr::summarizes trips on
the origin end by mode and region of origin.
}
