% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floorspace.R
\name{extract_volume}
\alias{extract_volume}
\title{Extract floorspace volume}
\usage{
extract_volume(...)
}
\arguments{
\item{...}{Arguments passed on to \code{extract_floorspace}
\describe{
  \item{db}{The scenario database.}
  \item{facet_var}{The variable in the zone table to facet by. Defaults to MPO}
  \item{facet_levels}{The levels of the facet variable to keep. Defaults to all
levels other than external stations.}
  \item{type_levels}{The types of employment to show in the plot.}
  \item{index}{Should the function extract indexed or absolute values?}
}}
}
\value{
A data frame with the total dollars sold.
}
\description{
This function extracts the total value of floorspace sold in each region by
floor_type.
}

