% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone_leaflet.R
\name{extract_zones}
\alias{extract_zones}
\title{Get zones geometry from the database}
\usage{
extract_zones(db, tbl_name = "ALLZONES", wkt_col = "WKTSURFACE",
  proj4string = "+proj=lcc +lat_1=43 +lat_2=45.5 +lat_0=41.75 +lon_0=-120.5 +x_0=399999.9999999999 +y_0=0 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=ft +no_defs")
}
\arguments{
\item{db}{class \code{src}.  The connection to the database.}

\item{tbl_name}{chr.  The name of the table in \code{db} containing the geometry
Defaults to ALLZONES.}

\item{wkt_col}{chr.  The name of the column in \code{tbl_name} containing the
WKT geometry. Defaults to WKTSURFACE}

\item{proj4string}{chr.  The proj4 string for the
\code{wkt_col} coordinates.  Defaults to Oregon Lambert.}
}
\value{
a \code{SpatialPolygonsDataFrame} in WGS84 lat long coordinates.
}
\description{
Get zones geometry from the database
}
\details{
This function extracts the geography for the zones from the scenario
         database. It automatically reprojects the data to WGS84 using \code{st_transform}.
}
