% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_validation.R
\name{get_validation_table}
\alias{get_validation_table}
\title{Get comparison of link volumes to ATR count}
\usage{
get_validation_table(db, year = c(2010, 2013), trucks = FALSE)
}
\arguments{
\item{db}{The scenario database.}

\item{year}{the year to build the comparison table for.}
}
\value{
A ggplot2 object.
}
\description{
Get comparison of link volumes to ATR count
}

