% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{links}
\alias{links}
\title{Link table}
\format{\preformatted{Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	218656 obs. of  28 variables:
 $ long      : num  -123 -123 -123 -123 -123 ...
 $ lat       : num  42.4 42.4 42.4 42.4 42.4 ...
 $ order     : int  1 2 1 2 1 2 1 2 1 2 ...
 $ piece     : Factor w/ 1 level "1": 1 1 1 1 1 1 1 1 1 1 ...
 $ group     : Factor w/ 25587 levels "0.1","1.1","2.1",..: 1 1 2 2 3 3 4 4 5 5 ...
 $ id        : chr  "0" "0" "1" "1" ...
 $ NO        : int  1 1 2 2 3 3 4 4 6 6 ...
 $ FROMNODENO: int  13235 13235 13239 13239 13241 13241 12360 12360 12371 12371 ...
 $ TONODENO  : int  13236 13236 13240 13240 13242 13242 12371 12371 12647 12647 ...
 $ TYPENO    : Factor w/ 2 levels "0","1": 1 1 1 1 1 1 1 1 1 1 ...
 $ TSYSSET   : Factor w/ 17 levels "a,b,d,e,f,g,h",..: 7 7 7 7 2 2 7 7 7 7 ...
 $ LENGTH    : Factor w/ 3243 levels "0.000mi","0.001mi",..: 160 160 150 150 127 127 53 53 133 133 ...
 $ BEG_MP    : int  0 0 0 0 0 0 0 0 0 0 ...
 $ END_MP    : int  0 0 0 0 0 0 0 0 0 0 ...
 $ FCTYPE    : Factor w/ 10 levels "Rural Freeway Ramp",..: NA NA NA NA NA NA NA NA NA NA ...
 $ FEDURBAN  : int  1 1 1 1 1 1 1 1 1 1 ...
 $ STHWY     : int  0 0 0 0 0 0 0 0 0 0 ...
 $ R_NO      : int  1 1 2 2 3 3 4 4 6 6 ...
 $ R_FROMNO.1: int  13236 13236 13240 13240 13242 13242 12371 12371 12647 12647 ...
 $ R_TONODENO: int  13235 13235 13239 13239 13241 13241 12360 12360 12371 12371 ...
 $ R_TYPENO  : Factor w/ 2 levels "0","1": 1 1 1 1 1 1 1 1 1 1 ...
 $ R_TSYSSET : Factor w/ 17 levels "a,b,d,e,f,g,h",..: 7 7 7 7 2 2 7 7 7 7 ...
 $ R_LENGTH  : Factor w/ 3241 levels "0.000mi","0.001mi",..: 160 160 150 150 127 127 53 53 133 133 ...
 $ R_BEG_MP  : int  0 0 0 0 0 0 0 0 0 0 ...
 $ R_END_MP  : int  0 0 0 0 0 0 0 0 0 0 ...
 $ R_FCTYPE  : Factor w/ 10 levels "Rural Freeway Ramp",..: NA NA NA NA NA NA NA NA NA NA ...
 $ R_FEDURBAN: int  1 1 1 1 1 1 1 1 1 1 ...
 $ R_STHWY   : int  0 0 0 0 0 0 0 0 0 0 ...
}}
\usage{
links
}
\description{
A fortified shapefile containing link attributes
}
\keyword{datasets}

