% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{links}
\alias{links}
\title{Link table}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 507 rows and 8 columns.}
\usage{
links
}
\description{
A fortified shapefile containing link attributes
}
\keyword{datasets}

