% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/employment.R
\name{multiple_employment}
\alias{multiple_employment}
\title{Compare Employment by sector across multiple scenarios.}
\usage{
multiple_employment(dbset, db_names, facet_var = c("BZONE", "MPO", "COUNTY",
  "STATE"), facet_levels = NULL, type_levels = NULL)
}
\arguments{
\item{dbset}{A list of connections to SWIM databases.}

\item{db_names}{A character vector naming the scenarios.}

\item{facet_var}{The variable in the zone table to facet by. Defaults to MPO}

\item{facet_levels}{The levels of the facet variable to keep. Defaults to all
levels other than external stations.}

\item{type_levels}{The types of employment to show in the plot.}
}
\value{
a ggplot2 object.
}
\description{
Compare Employment by sector across multiple scenarios.
}

