% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floorspace.R
\name{multiple_floorspace}
\alias{multiple_floorspace}
\title{Compare floorspace by type across multiple scenarios.}
\usage{
multiple_floorspace(dbset, db_names, ..., variable = c("floorspace", "rent",
  "occupancy"))
}
\arguments{
\item{dbset}{A list of connections to SWIM databases.}

\item{db_names}{A character vector naming the scenarios.}

\item{...}{Arguments passed on to \code{extract_floorspace}
\describe{
  \item{db}{The scenario database.}
  \item{facet_var}{The variable in the zone table to facet by. Defaults to MPO}
  \item{facet_levels}{The levels of the facet variable to keep. Defaults to all
levels other than external stations.}
  \item{type_levels}{The types of employment to show in the plot.}
  \item{index}{Should the function extract indexed or absolute values?}
}}

\item{variable}{The variable to plot, one of rents, occupancy rate, or new
floorspace}
}
\value{
a ggplot2 object.
}
\description{
Compare floorspace by type across multiple scenarios.
}

