% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/floorspace.R
\name{multiple_floorspace}
\alias{multiple_floorspace}
\title{Compare floorspace by type across multiple scenarios.}
\usage{
multiple_floorspace(dbset, db_names, facet_var = c("MPO", "COUNTY", "STATE"),
  facet_levels = NULL, type_levels = NULL)
}
\arguments{
\item{dbset}{A list of connections to SWIM databases.}

\item{db_names}{A character vector naming the scenarios.}

\item{facet_var}{The variable in the zone table to facet by. Defaults to MPO}

\item{facet_levels}{The levels of the facet variable to keep. Defaults to all
levels other than external stations.}

\item{type_levels}{The types of floorspace to show in the plot.}
}
\value{
a ggplot2 object.
}
\description{
Compare floorspace by type across multiple scenarios.
}

