% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{multiple_sevar}
\alias{multiple_sevar}
\title{Compare Population or Employment across multiple scenarios.}
\usage{
multiple_sevar(dbset, db_names, variable = c("population", "employment"), ...)
}
\arguments{
\item{dbset}{A list of connections to SWIM databases.}

\item{db_names}{A character vector naming the scenarios.}

\item{variable}{One of \code{c("population", "employment")} defining which
socioeconomic variable to include in}

\item{...}{Arguments passed on to \code{extract_se}
\describe{
  \item{db}{The scenario sqlite database.}
  \item{color_var}{Field to color by: either "MPO" or "COUNTY".}
  \item{color_levels}{A character vector of the color variable specifiying
which levels to include.}
  \item{controls}{Plot against the control totals. Defaults to TRUE, cannot
currently run with FALSE.}
  \item{index}{Whether to show the variables as indexed against the base year.}
}}
}
\value{
a ggplot2 object.
}
\description{
Compare Population or Employment across multiple scenarios.
}

