% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multiplots.R
\name{multiple_sevar}
\alias{multiple_sevar}
\title{Compare Population or Employment across multiple scenarios.}
\usage{
multiple_sevar(dbset, db_names, variable = c("population", "employment"),
  facet_var = c("MPO", "COUNTY"), facet_levels = NULL)
}
\arguments{
\item{dbset}{A list of connections to SWIM databases.}

\item{db_names}{A character vector naming the scenarios.}

\item{variable}{One of \code{c("population", "employment")} defining which
socioeconomic variable to include in}

\item{facet_var}{Field to facet by: either "MPO" or "COUNTY".}

\item{facet_levels}{A character vector of the facet variable specifiying
  which levels to include.}
}
\value{
a ggplot2 object.
}
\description{
Compare Population or Employment across multiple scenarios.
}

