% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trips.R
\name{multiple_trips}
\alias{multiple_trips}
\title{Plot trips in multiple scenarios.}
\usage{
multiple_trips(dbset, db_names, ...)
}
\arguments{
\item{dbset}{A list of connections to SWIM databases.}

\item{db_names}{A character vector naming the scenarios.}

\item{...}{Arguments passed on to \code{extract_trips}
\describe{
  \item{db}{the scenario database.}
  \item{facet_var}{The region to dplyr::summarize by.}
  \item{facet_levels}{Regions to include in summary.}
  \item{color_levels}{Modes to include in summary. Defaults to all modes other
than \code{school}. See consolidated modes in \link{mode_types}.}
  \item{index}{whether to index mode split off the base year.}
}}
}
\value{
A ggplot2 object
}
\description{
Plot trips in multiple scenarios.
}

