% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trips.R
\name{multiple_trips}
\alias{multiple_trips}
\title{Plot trips in multiple scenarios.}
\usage{
multiple_trips(dbset, db_names, facet_var = c("MPO", "COUNTY", "STATE"),
  facet_levels = NULL, color_levels = c("auto", "transit", "non-motorized",
  "truck"))
}
\arguments{
\item{dbset}{A list of connections to SWIM databases.}

\item{db_names}{A character vector naming the scenarios.}

\item{facet_var}{The region to dplyr::summarize by.}

\item{facet_levels}{Regions to include in summary.}

\item{color_levels}{Modes to include in summary. Defaults to all modes other
than \code{school}. See consolidated modes in \link{mode_types}.}
}
\description{
Plot trips in multiple scenarios.
}

