% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/laborforce.R
\name{multiple_wapr}
\alias{multiple_wapr}
\title{Compare WAPR  across multiple scenarios.}
\usage{
multiple_wapr(dbset, db_names, facet_var = c("BZONE", "MPO", "COUNTY",
  "STATE"), facet_levels = NULL)
}
\arguments{
\item{dbset}{A list of connections to SWIM databases.}

\item{db_names}{A character vector naming the scenarios.}

\item{facet_levels}{A character vector of the facet variable specifiying
  which levels to include.}
}
\value{
a ggplot2 object.
}
\description{
Compare WAPR  across multiple scenarios.
}

