% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_validation.R
\name{plot_countcomparison}
\alias{plot_countcomparison}
\title{Plot base-year traffic validation.}
\usage{
plot_countcomparison(db, year = c(2010, 2013, 2015), trucks = FALSE)
}
\arguments{
\item{db}{The scenario database.}

\item{year}{the year to use in the validation plot.}

\item{trucks}{Build a truck count validation table? Defaults to FALSE}
}
\value{
A ggplot2 object.
}
\description{
Create a plot of the base scenario link volumes compared with observed AADT.
}

