% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employment.R
\name{plot_employment}
\alias{plot_employment}
\title{Plot Employment by Sector}
\usage{
plot_employment(db, ...)
}
\arguments{
\item{...}{Arguments passed on to \code{extract_employment}
\describe{
  \item{db}{The scenario database.}
  \item{facet_var}{The variable in the zone table to facet by. Defaults to MPO}
  \item{facet_levels}{The levels of the facet variable to keep. Defaults to all
levels other than external stations.}
  \item{type_levels}{The types of employment to show in the plot.}
  \item{employment_categories}{[Optional] data_frame of employment aggregation categories;
if \code{NULL}, uses \code{\link{emp_types}}. Should have columns
\code{ACTIVITY} (full code of activity in AA databases) and \code{emp_type}.}
}}
}
\value{
A ggplot2 object showing the employment by type and and year.
}
\description{
This function plots the employment by type in an area over time between two
scenarios.
}
