% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floorspace.R
\name{plot_floorspace}
\alias{plot_floorspace}
\title{Plot floorspace statistics over time}
\usage{
plot_floorspace(db, facet_var = c("MPO", "COUNTY", "STATE"),
  facet_levels = NULL, type_levels = NULL, price = FALSE)
}
\arguments{
\item{db}{The scenario database.}

\item{facet_var}{The variable in the zone table to facet by. Defaults to MPO}

\item{facet_levels}{The levels of the facet variable to keep. Defaults to all
levels other than external stations.}

\item{type_levels}{The types of floorspace to show in the plot.}

\item{price}{Print price instead of floorspace.}
}
\value{
A ggplot2 object showing the floorspace by type and and year.
}
\description{
This function plots the constructed floorspace by type over time facetted by a
variable chosen from the zone attributes table.
}

