% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdp.R
\name{plot_gdp}
\alias{plot_gdp}
\title{Plot the GDP contribution of Labor over time in a scenario}
\usage{
plot_gdp(db, facet_var = c("MPO", "COUNTY", "STATE"), facet_levels = NULL,
  color_levels = NULL)
}
\arguments{
\item{db}{The scenario sqlite database.}

\item{facet_var}{Field to facet by: "MPO", "COUNTY", or "STATE".}

\item{facet_levels}{A character vector of the variable specifiying
which levels to include.}

\item{color_levels}{A character vector of the industry sectors to include.
Defaults to all.}
}
\value{
a ggplot2 plot object.
}
\description{
Plot the GDP contribution of Labor over time in a scenario
}

