% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logsums.R
\name{plot_logsums}
\alias{plot_logsums}
\title{Plot average logsums over time for a single scenario}
\usage{
plot_logsums(db, color_var = c("COUNTY", "MPO", "ALDREGION", "STATE",
  "DOT_REGION"), color_levels = NULL)
}
\arguments{
\item{db}{The scenario sqlite database.}

\item{color_var}{The variable to color the plot by.}

\item{color_levels}{The variable to}
}
\description{
Plot average logsums over time for a single scenario
}
