% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floorspace.R
\name{plot_occupancy}
\alias{plot_occupancy}
\title{Plot occupancy in a scenario over time.}
\usage{
plot_occupancy(...)
}
\arguments{
\item{...}{Arguments passed on to \code{extract_rents}
\describe{
  \item{db}{The scenario database.}
  \item{facet_var}{The variable in the zone table to facet by. Defaults to MPO}
  \item{facet_levels}{The levels of the facet variable to keep. Defaults to all
levels other than external stations.}
  \item{type_levels}{The types of employment to show in the plot.}
  \item{index}{Should the function extract indexed or absolute values?}
}}
}
\description{
Plot occupancy in a scenario over time.
}

