% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmt.R
\name{plot_pct_cong}
\alias{plot_pct_cong}
\title{Plot percent congested links over time.}
\usage{
plot_pct_cong(db, facet_var = "MPO", facet_levels = NULL,
  congested_voc = 0.9)
}
\arguments{
\item{db}{The scenario database.}

\item{facet_var}{The variable to use in faceting the data.}

\item{facet_levels}{A character vector giving the levels of the faceting
variable to use. Levels not called are dropped from the plot; default is
\code{NULL}, meaning print all level.}

\item{congested_voc}{the volume to capacity ratio considered as "congested."
Default value is \code{0.9}.}
}
\value{
A ggplot2 figure object.
}
\description{
This function plots the percent of PM VMT traveling on congested links.,
faceted by arbitrary variables in the link table.
}

