% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_sevar}
\alias{plot_sevar}
\title{Make a plot of population and employment}
\usage{
plot_sevar(df, which_county, controls = TRUE)
}
\arguments{
\item{df}{A \code{data_frame} consisting of at least the following variables:
  county, year, population, employment.}

\item{which_county}{A character vector of county names whose values should
  be included in the plot.}

\item{controls}{Plot against the control totals. Defaults to TRUE, cannot
  currently run with FALSE.}
}
\value{
A \code{ggplot2} plot object showing the modeled change in employment
  and population over time.
}
\description{
Make a plot of population and employment
}

