% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{plot_sevar}
\alias{plot_sevar}
\title{Make a plot of population and employment}
\usage{
plot_sevar(...)
}
\arguments{
\item{...}{Arguments passed on to \code{extract_se}
\describe{
  \item{db}{The scenario sqlite database.}
  \item{color_var}{Field to color by: either "MPO" or "COUNTY".}
  \item{color_levels}{A character vector of the color variable specifiying
which levels to include.}
  \item{controls}{Plot against the control totals. Defaults to TRUE, cannot
currently run with FALSE.}
  \item{index}{Whether to show the variables as indexed against the base year.}
}}
}
\value{
A \code{ggplot2} plot object showing the modeled change in employment
  and population over time.
}
\description{
Make a plot of population and employment
}
