% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlfd.R
\name{plot_tlfd}
\alias{plot_tlfd}
\title{Plot a trip length frequency distribution}
\usage{
plot_tlfd(db, region_var = c("MPO", "COUNTY", "STATE"), regions = NULL,
  bin_width = 1, max_bin = 50)
}
\arguments{
\item{db}{The scenario database.}

\item{region_var}{The region to aggregate to.}

\item{regions}{The regions to return.  If \code{NULL}, returns all.}

\item{bin_width}{The width of bins to use in the trip length frequency
distribution. Defaults to 1.}

\item{max_bin}{The maximum bin value. Defaults to 50.}
}
\value{
A ggplot2 object.
}
\description{
Plot a trip length frequency distribution
}

