% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trips.R
\name{plot_trips}
\alias{plot_trips}
\title{Plot Trip productions}
\usage{
plot_trips(..., share = TRUE)
}
\arguments{
\item{...}{Arguments passed on to \code{extract_trips}
\describe{
  \item{db}{the scenario database.}
  \item{facet_var}{The region to dplyr::summarize by.}
  \item{facet_levels}{Regions to include in summary.}
  \item{color_levels}{Modes to include in summary. Defaults to all modes other
than \code{school}. See consolidated modes in \link{mode_types}.}
  \item{index}{whether to index mode split off the base year.}
}}

\item{share}{Plot mode share instead of total trips? Defaults to \code{TRUE}.}
}
\value{
A ggplot2 object.
}
\description{
Plot Trip productions
}

