% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_validation.R
\name{plot_validation}
\alias{plot_validation}
\title{Plot base-year traffic validation.}
\usage{
plot_validation(db, facet_var = "MPO", facet_levels = NULL)
}
\description{
Plot base-year traffic validation.
}

