% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmt.R
\name{plot_vht}
\alias{plot_vht}
\title{Plot VHT over time.}
\usage{
plot_vht(db, facet_var = "MPO", facet_levels = NULL)
}
\arguments{
\item{db}{The scenario database.}

\item{facet_var}{The variable to use in faceting the data.}

\item{facet_levels}{A character vector giving the levels of the faceting
variable to use. Levels not called are dropped from the plot; default is
\code{NULL}, meaning print all level.}
}
\value{
A ggplot2 figure object.
}
\description{
This function plots the sum of the daily vehicle travel time, colored and
faceted by arbitrary variables in the link or zone table.
}
