% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_counts}
\alias{ref_counts}
\title{Reference counts}
\format{A data frame with 2,170 observations.
\describe{
  \item{site}{The site location of the count.}
  \item{FROMNODENO}{The from node of the link in the SWIM highway network.}
  \item{TONODENO}{The from node of the link in the SWIM highway network.}
  \item{aawdt}{Observed traffic flow; calculated AAWDT by ODOT data analysis
  division.}
  \item{year}{The year of the count.}
}}
\source{
Calculated by Rick Donnelly from data supplied by ODOT.
}
\usage{
ref_counts
}
\description{
A dataset containing the reference highway counts for comparing modeled
traffic volumes in multiple historical years.
}
\keyword{datasets}

