% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_times.R
\name{report_top_times}
\alias{report_top_times}
\title{Query Database for Event Rankings}
\usage{
report_top_times(con, conference = "All", team_name = "All",
  gender = "Both", athlete = NULL, event = "All", top = 3)
}
\arguments{
\item{con}{a database connection.}

\item{conference}{a character vector with the name of an NCAA swimming conference.  If not provided the output include all conferences.}

\item{team_name}{A character vector with the name of an NCAA swimming conference.  If not provided the output include all teams.}

\item{gender}{a character vector.  Indicates the output should include men, women or both.  Defaults to "Both".  Other options are "M" for men and "F" for women.}

\item{athlete}{a character vector with the name or athlete id of a swimmier.  If not provided the results include all swimmers.  The athlete id follows the USA Swimming standard.}

\item{event}{a character vector with the name of swimming envent.  If not provided the results include all events.}

\item{top}{an integer indicating how many results to return per event  When 
queryin for an individual swimmer it only returns 1 result regardless of 
the value provided .}
}
\value{
a data frame of ranked swimming results.
}
\description{
Extract Top Times
}
\details{
\code{report_top_times} queries a SQL database to produce a ranking of swimmers in the specified events.  The function is flexiable and can rank swimmers across all of Division 1, within a specific conference or within a team.  It can also produce an individual swimmer's best result in each event they have competed in.
}
