% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim_display.R
\name{skim_popup}
\alias{skim_popup}
\title{Build skim popup tag for leaflet}
\usage{
skim_popup(shp, zone)
}
\arguments{
\item{shp}{A shapefile containing data table attributes}

\item{zone}{The focus zone.}
}
\description{
Build skim popup tag for leaflet
}

