% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swim_db_data_load.R
\name{swim_db_data_load}
\alias{swim_db_data_load}
\title{Load college swimming results into a database.}
\usage{
swim_db_data_load(con, data, conference_name)
}
\arguments{
\item{con}{A SQLiteConnection object.}

\item{data}{A data.frame containing swimming results.  The data.frame should follow the form of a USA Swimming Top Times report and been cleaned by \code{\link{clean_swim()}}.}

\item{conference_name}{A character string providing the name of the conference repersented in /emph{data}.}
}
\description{
Load college swimming results into a database.
}
\details{
\code{swim_db_data_load()} allows users to easily load swimming results data into a  database.  It is specifically designed to load data from from USA Swimmings Top Times Report.  Conference data, team data, athlete data, meet data and result data is loaded into the database with the function.

The function is set up so only distinct data is stored.  If a the function encounters a duplicate record it skips the records and moves on to the next record.
}
