% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swimR-package.r
\docType{package}
\name{swimR}
\alias{swimR}
\alias{swimR-package}
\title{swimR: A package for collection of college swimming resutls.}
\description{
swimR provides functions to enable the collection and analysis of college swimming results from \href{http://www.usaswimming.org/DesktopDefault.aspx?TabId=1971&Alias=Rainbow&Lang=en}{USA Swimming's Top Times} report.
}
\details{
The package facilitates automated data collection of college swimming results by utlizling \code{\link[RSelenium]{RSelenium}}.  In addition to retreving the data the package provides basic functions to clean the colected data.  It also has functions to store the data in a SQLite database.

Overtime we hope to expand the capabilities of the package.
}

