% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zones}
\alias{zones}
\title{Zone table}
\format{\preformatted{Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	701344 obs. of  12 variables:
 $ long     : num  -123 -123 -123 -123 -123 ...
 $ lat      : num  45.5 45.5 45.5 45.5 45.5 ...
 $ order    : int  1 2 3 4 5 6 7 8 9 10 ...
 $ hole     : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ piece    : Factor w/ 3 levels "1","2","3": 1 1 1 1 1 1 1 1 1 1 ...
 $ group    : Factor w/ 2991 levels "0.1","1.1","2.1",..: 1 1 1 1 1 1 1 1 1 1 ...
 $ id       : chr  "0" "0" "0" "0" ...
 $ NO       : int  NA NA NA NA NA NA NA NA NA NA ...
 $ ALDREGION: Factor w/ 15 levels "Astoria_LincolnCity",..: NA NA NA NA NA NA NA NA NA NA ...
 $ AZONE    : Factor w/ 2986 levels "1","10","100",..: NA NA NA NA NA NA NA NA NA NA ...
 $ BZONE    : int  NA NA NA NA NA NA NA NA NA NA ...
 $ AREASQFT : num  NA NA NA NA NA NA NA NA NA NA ...
}}
\usage{
zones
}
\description{
A fortified shapefile containing the zone geometries
}
\keyword{datasets}

